+++ client/src/io/worxbend/gitea4s/http/GiteaEndpoint.scala
@@ -414,6 +414,20 @@ object GiteaEndpoints:
       response = "#/responses/ContentsResponse"
     )
 
+  val repoGetArchive: GiteaEndpoint =
+    GiteaEndpoint(
+      method = "GET",
+      path = "/repos/{owner}/{repo}/archive/{archive}",
+      operationId = "repoGetArchive",
+      parameters = List(
+        GiteaParameter("owner", "path", required = true),
+        GiteaParameter("repo", "path", required = true),
+        GiteaParameter("archive", "path", required = true),
+        GiteaParameter("path", "query", required = false)
+      ),
+      response = "description: success"
+    )
+
   val repoGetRawFile: GiteaEndpoint =
     GiteaEndpoint(
       method = "GET",
@@ -1443,6 +1457,42 @@ object GiteaEndpoints:
         GiteaParameter("repo", "path", required = true),
         GiteaParameter("index", "path", required = true),
         GiteaParameter("id", "path", required = true)
+      ),
+      response = "#/responses/empty"
+    )
+
+  val createCurrentUserRepo: GiteaEndpoint =
+    GiteaEndpoint(
+      method = "POST",
+      path = "/user/repos",
+      operationId = "createCurrentUserRepo",
+      parameters = List(
+        GiteaParameter("body", "body", required = false)
+      ),
+      response = "#/responses/Repository"
+    )
+
+  val createFork: GiteaEndpoint =
+    GiteaEndpoint(
+      method = "POST",
+      path = "/repos/{owner}/{repo}/forks",
+      operationId = "createFork",
+      parameters = List(
+        GiteaParameter("owner", "path", required = true),
+        GiteaParameter("repo", "path", required = true),
+        GiteaParameter("body", "body", required = false)
+      ),
+      response = "#/responses/Repository"
+    )
+
+  val repoDelete: GiteaEndpoint =
+    GiteaEndpoint(
+      method = "DELETE",
+      path = "/repos/{owner}/{repo}",
+      operationId = "repoDelete",
+      parameters = List(
+        GiteaParameter("owner", "path", required = true),
+        GiteaParameter("repo", "path", required = true)
       ),
       response = "#/responses/empty"
     )