+++ client/src/io/worxbend/gitea4s/internal/SttpGiteaClient.scala
@@ -39,6 +39,7 @@ import io.worxbend.gitea4s.model.{
   ContentsResponse,
   CreatePullRequestOption,
   CreatePullReviewOptions,
+  CreateRepo,
   CreateStatusOption,
   DismissPullReviewOptions,
   EditDeadlineOption,
@@ -46,6 +47,7 @@ import io.worxbend.gitea4s.model.{
   EditIssue,
   EditPullRequestOption,
   EditReactionOption,
+  ForkRepo,
   GitBlobResponse,
   GitTreeResponse,
   Issue,
@@ -206,6 +208,18 @@ final class SttpGiteaClient(config: GiteaConfig, backend: Backend[Task]) extends
       params: ContentsParams
   ): IO[GiteaError, Chunk[Byte]] =
     executor.send(GiteaRequests.repoMediaFile(config, owner, repo, filepath, params))
+
+  override def archive(owner: String, repo: String, archive: String): IO[GiteaError, Chunk[Byte]] =
+    executor.send(GiteaRequests.repoGetArchive(config, owner, repo, archive))
+
+  override def create(body: CreateRepo): IO[GiteaError, Repository] =
+    executor.send(GiteaRequests.createRepo(config, body))
+
+  override def fork(owner: String, repo: String, body: ForkRepo): IO[GiteaError, Repository] =
+    executor.send(GiteaRequests.forkRepo(config, owner, repo, body))
+
+  override def delete(owner: String, repo: String): IO[GiteaError, Unit] =
+    executor.send(GiteaRequests.deleteRepo(config, owner, repo))
 
   override def list(
       owner: String,